backend/requirements.txt
# Flask Backend
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-Migrate>=4.0.0
Flask-RESTX>=1.3.0
Flask-Limiter>=3.7.0
Flask-SQLAlchemy>=3.1.0
Werkzeug>=3.0.0
gunicorn>=22.0.0

# Database
SQLAlchemy>=2.0.16
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Machine Learning & AI
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.17.3,<2.0
pandas>=1.3.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
